requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
sse-starlette==3.3.4
pydantic==2.13.3
google-genai==1.73.1
git+https://github.com/stepacool/mcphero.git@bbc1ba2c6039fa9ac51c95daec8b3f497c31d646
